requirements.txt
asgiref==3.7.2
astroid==2.15.8
attrs==23.1.0
coverage==7.3.1
dill==0.3.7
Django==4.1.6
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
drf-spectacular==0.26.4
exceptiongroup==1.1.3
filetype==1.2.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
phonenumberslite==8.13.22
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
PyJWT==2.8.0
pylint==2.17.7
pytest==7.4.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
wrapt==1.15.0
